fix(vision): widen PNG chunk-length bounds math to 64-bit (review cycle 2)
Belt-and-suspenders: (long)dataOff + clen + 4 so a chunk length near
int.MaxValue (only reachable with a ~2 GB in-memory PNG) can't wrap
negative and defeat the truncation check. Closes the last review nit.

+++ src/SharpInference.Vision/ImageIO.cs
@@ -45,7 +45,7 @@ public static class ImageIO
             int dataOff = pos + 8;
             // ReadUInt32BE returns a signed int; a high-bit-set length is negative and would
             // wrap past the bounds check, so reject it explicitly (malformed/hostile PNG).
-            if (clen < 0 || dataOff + clen + 4 > len)
+            if (clen < 0 || (long)dataOff + clen + 4 > len)
                 throw new InvalidDataException($"Truncated PNG chunk '{type}'.");
 
             switch (type)
